+++ src/main/java/com/nexaworks/rafiq/rabbit/publisher/OTPEmailPublisher.java
@@ -4,8 +4,8 @@ import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
-import com.nexaworks.rafiq.constant.RabbitMQConstant;
-import com.nexaworks.rafiq.dto.notificaiton.EmailNotification;
+import com.nexaworks.rafiq.rabbit.constant.RabbitMQConstant;
+import com.nexaworks.rafiq.rabbit.notificaiton.EmailNotification;
 
 import lombok.extern.slf4j.Slf4j;
